pyproject.toml
[project]
name = "medical-dashboard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
